pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Jickx <valentin.sharonov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.3"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.7"
python-dotenv = "^1.0.0"
validators = "^0.21.2"
url-normalize = "^1.4.3"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
